Restructured the project, added util to validate input quantity
