+++ src/main/java/com/es/phoneshop/dao/ArrayListProductDao.java
@@ -1,4 +1,10 @@
-package com.es.phoneshop.model.product;
+package com.es.phoneshop.dao;
+
+import com.es.phoneshop.enumeration.SortField;
+import com.es.phoneshop.enumeration.SortOrder;
+import com.es.phoneshop.exception.ProductNotFoundException;
+import com.es.phoneshop.model.product.PriceHistory;
+import com.es.phoneshop.model.product.Product;
 
 import java.time.LocalDateTime;
 import java.util.*;
