+++ src/main/java/com/es/phoneshop/enumeration/SortOrder.java
@@ -1,4 +1,4 @@
-package com.es.phoneshop.model.product;
+package com.es.phoneshop.enumeration;
 
 public enum SortOrder {
     asc, desc
