+++ src/main/java/com/es/phoneshop/service/CartService.java
@@ -1,4 +1,7 @@
-package com.es.phoneshop.model.cart;
+package com.es.phoneshop.service;
+
+import com.es.phoneshop.exception.OutOfStockException;
+import com.es.phoneshop.model.cart.Cart;
 
 import javax.servlet.http.HttpServletRequest;
