+++ src/main/java/com/es/phoneshop/service/DefaultRecentlyViewedProductsService.java
@@ -1,8 +1,9 @@
-package com.es.phoneshop.model.recentlyViewedProducts;
+package com.es.phoneshop.service;
 
-import com.es.phoneshop.model.product.ArrayListProductDao;
+import com.es.phoneshop.dao.ArrayListProductDao;
 import com.es.phoneshop.model.product.Product;
-import com.es.phoneshop.model.product.ProductDao;
+import com.es.phoneshop.dao.ProductDao;
+import com.es.phoneshop.model.recentlyViewedProducts.RecentlyViewedProducts;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
