+++ src/main/java/com/es/phoneshop/service/RecentlyViewedProductsService.java
@@ -1,4 +1,6 @@
-package com.es.phoneshop.model.recentlyViewedProducts;
+package com.es.phoneshop.service;
+
+import com.es.phoneshop.model.recentlyViewedProducts.RecentlyViewedProducts;
 
 import javax.servlet.http.HttpServletRequest;
