+++ src/main/webapp/WEB-INF/pages/cart.jsp
@@ -18,7 +18,7 @@
         </c:if>
         <c:if test="${not empty errors}">
             <div class="error">
-                There was an error added to cart
+                There was an error while updating the cart
             </div>
         </c:if>
         <form method="post" action="${pageContext.servletContext.contextPath}/cart">
