+++ src/main/webapp/WEB-INF/pages/product.jsp
@@ -16,7 +16,7 @@
     </c:if>
     <c:if test="${not empty error}">
         <div class="error">
-            There was an error added to cart
+            There was an error while adding to the cart
         </div>
     </c:if>
     <h1>
