+++ src/test/java/com/es/phoneshop/model/product/ArrayListProductDaoTest.java
@@ -1,5 +1,10 @@
 package com.es.phoneshop.model.product;
 
+import com.es.phoneshop.dao.ArrayListProductDao;
+import com.es.phoneshop.dao.ProductDao;
+import com.es.phoneshop.enumeration.SortField;
+import com.es.phoneshop.enumeration.SortOrder;
+import com.es.phoneshop.exception.ProductNotFoundException;
 import org.junit.Before;
 import org.junit.Test;
