+++ src/test/java/com/es/phoneshop/web/servlet/ProductDetailsPageServletTest.java
@@ -1,11 +1,11 @@
-package com.es.phoneshop.web;
+package com.es.phoneshop.web.servlet;
 
+import com.es.phoneshop.dao.ArrayListProductDao;
 import com.es.phoneshop.model.cart.Cart;
-import com.es.phoneshop.model.cart.CartService;
-import com.es.phoneshop.model.product.ArrayListProductDao;
 import com.es.phoneshop.model.product.Product;
 import com.es.phoneshop.model.recentlyViewedProducts.RecentlyViewedProducts;
-import com.es.phoneshop.model.recentlyViewedProducts.RecentlyViewedProductsService;
+import com.es.phoneshop.service.CartService;
+import com.es.phoneshop.service.RecentlyViewedProductsService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
